Move flags to constrained file
